Don't cast IfdFormat in TagValue::Invalid
There's no reason to discard it, and the semantics could be useful.

+++ src/ifdformat.rs
@@ -43,7 +43,7 @@ pub fn tag_value_new(f: &IfdEntry) -> TagValue
 		IfdFormat::U16 => {
 			if f.data.len() < (f.count as usize * 2) {
 				return TagValue::Invalid(f.data.clone(), f.le,
-							             f.format as u16, f.count);
+							             f.format, f.count);
 			}
 			let a = read_u16_array(f.le, f.count, &f.data[..]);
 			TagValue::U16(a)
@@ -51,7 +51,7 @@ pub fn tag_value_new(f: &IfdEntry) -> TagValue
 		IfdFormat::I16 => {
 			if f.data.len() < (f.count as usize * 2) {
 				return TagValue::Invalid(f.data.clone(), f.le,
-							             f.format as u16, f.count);
+							             f.format, f.count);
 			}
 			let a = read_i16_array(f.le, f.count, &f.data[..]);
 			TagValue::I16(a)
@@ -59,7 +59,7 @@ pub fn tag_value_new(f: &IfdEntry) -> TagValue
 		IfdFormat::U8 => {
 			if f.data.len() < (f.count as usize * 1) {
 				return TagValue::Invalid(f.data.clone(), f.le,
-							             f.format as u16, f.count);
+							             f.format, f.count);
 			}
 			let a = f.data.clone();
 			TagValue::U8(a)
@@ -67,7 +67,7 @@ pub fn tag_value_new(f: &IfdEntry) -> TagValue
 		IfdFormat::I8 => {
 			if f.data.len() < (f.count as usize * 1) {
 				return TagValue::Invalid(f.data.clone(), f.le,
-							             f.format as u16, f.count);
+							             f.format, f.count);
 			}
 			let a = read_i8_array(f.count, &f.data[..]);
 			TagValue::I8(a)
@@ -75,7 +75,7 @@ pub fn tag_value_new(f: &IfdEntry) -> TagValue
 		IfdFormat::U32 => {
 			if f.data.len() < (f.count as usize * 4) {
 				return TagValue::Invalid(f.data.clone(), f.le,
-							             f.format as u16, f.count);
+							             f.format, f.count);
 			}
 			let a = read_u32_array(f.le, f.count, &f.data[..]);
 			TagValue::U32(a)
@@ -83,7 +83,7 @@ pub fn tag_value_new(f: &IfdEntry) -> TagValue
 		IfdFormat::I32 => {
 			if f.data.len() < (f.count as usize * 4) {
 				return TagValue::Invalid(f.data.clone(), f.le,
-							             f.format as u16, f.count);
+							             f.format, f.count);
 			}
 			let a = read_i32_array(f.le, f.count, &f.data[..]);
 			TagValue::I32(a)
@@ -91,7 +91,7 @@ pub fn tag_value_new(f: &IfdEntry) -> TagValue
 		IfdFormat::F32 => {
 			if f.data.len() < (f.count as usize * 4) {
 				return TagValue::Invalid(f.data.clone(), f.le,
-							             f.format as u16, f.count);
+							             f.format, f.count);
 			}
 			let a = read_f32_array(f.count, &f.data[..]);
 			TagValue::F32(a)
@@ -99,7 +99,7 @@ pub fn tag_value_new(f: &IfdEntry) -> TagValue
 		IfdFormat::F64 => {
 			if f.data.len() < (f.count as usize * 8) {
 				return TagValue::Invalid(f.data.clone(), f.le,
-							             f.format as u16, f.count);
+							             f.format, f.count);
 			}
 			let a = read_f64_array(f.count, &f.data[..]);
 			TagValue::F64(a)
@@ -107,7 +107,7 @@ pub fn tag_value_new(f: &IfdEntry) -> TagValue
 		IfdFormat::URational => {
 			if f.data.len() < (f.count as usize * 8) {
 				return TagValue::Invalid(f.data.clone(), f.le,
-							             f.format as u16, f.count);
+							             f.format, f.count);
 			}
 			let a = read_urational_array(f.le, f.count, &f.data[..]);
 			TagValue::URational(a)
@@ -115,7 +115,7 @@ pub fn tag_value_new(f: &IfdEntry) -> TagValue
 		IfdFormat::IRational => {
 			if f.data.len() < (f.count as usize * 8) {
 				return TagValue::Invalid(f.data.clone(), f.le,
-							             f.format as u16, f.count);
+							             f.format, f.count);
 			}
 			let a = read_irational_array(f.le, f.count, &f.data[..]);
 			TagValue::IRational(a)
@@ -129,4 +129,3 @@ pub fn tag_value_new(f: &IfdEntry) -> TagValue
 		_ => TagValue::Unknown(f.data.clone(), f.le)
 	}
 }
-
